cacheing and transformation of proposal list item
