+++ src/services/aragon-sdk/selectors/index.ts
@@ -1,0 +1,1 @@
+export * from './proposal';
